Fixed wrong centroid for square.
The centroid for both rectangle and square is now calculated based on the width and height from the top left corner instead of the start-point and endpoint.

This is because the endpoint stays the same when switching to square mode giving th esame centroid as a rectangle which is wrong. However, with width and height we only need to make them equal in the square case to calculate the correct centroid.