Used Memento pattern to fully implement undo/redo. It may have some bugs.
Now undo,redo works when moving shapes, rotating them or scaling them.

Essentially each time the user interacts with the drawing panel a new state is created by deep-copying the previous one.

A list of drawing panel states is used to provide undo/redo functionality.

The list of drawing panel states is essentially a list of lists of shapes.

More testing is needed to check if the undo, redo works correctly!